requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.2.0
tqdm>=4.66.0
PyYAML>=6.0.1
chardet>=5.2.0
pandas>=2.1.0
matplotlib>=3.8.0
pytest>=8.0.0
